// frontend/src/api/generated/models/createAdjustmentDtoType.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */

export type CreateAdjustmentDtoType = typeof CreateAdjustmentDtoType[keyof typeof CreateAdjustmentDtoType];


export const CreateAdjustmentDtoType = {
  destruction: 'destruction',
  mutation: 'mutation',
  swap: 'swap',
} as const;