// frontend/src/api/generated/models/itemDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */
import type { AttributeDto } from './attributeDto';
import type { AttributeOptionDto } from './attributeOptionDto';
import type { PictureDto } from './pictureDto';
import type { VariationDto } from './variationDto';

export interface ItemDto {
  id: string;
  title: string;
  categoryId: string;
  price: number;
  currencyId: string;
  initialQuantity: number;
  availableQuantity: number;
  soldQuantity: number;
  status: string;
  health?: number;
  logisticType: string;
  permalink: string;
  thumbnail: string;
  dateCreated: string;
  lastUpdated: string;
  startTime: string;
  stopTime: string;
  expirationTime: string;
  pictures: PictureDto[];
  attributes: AttributeDto[];
  attributeOptions: AttributeOptionDto[];
  variations: VariationDto[];
}